# doc/phase1/swagger.yaml
swagger: "2.0"
info:
  description: "This is the api spec for Team SpaceWhales."
  version: "1.0.0"
  title: ""
  contact: 
    name: ""
  license:
    name: "<insert license here>"
host: "<unknown_domain_insert_here>"
basePath: "/space_whales"
...
schemes:
- "https"
paths:
...
  /tracking/all:
    post:
      tags:
      - "Tracking"
      summary: "Get all product information for a user defined by their token"
      description: ""
      operationId: "getTrackedProducts"
      consumes: 
      - "application/json"
      produces: 
      - "application/json"
      parameters: 
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/LoginToken"
      responses:
        200:
          description: "Successful"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ProductItem"
        401:
          description: "Invalid/expired login token"
        405:
          description: "Product does not exist"
    
    put:
      tags: 
      - "Tracking"
      summary: "Add a product to the user's wishlist"
      description: ""
      operationId: "addTrackedProduct"
      consumes: 
      - "application/json"
      produces: 
      - "application/json"
      parameters: 
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NewProductRequest"
      responses:
        200:
          description: "Successful"
          schema:
            $ref: "#/definitions/ProductItem"
        401:
          description: "Invalid/expired login token"
        405:
          description: "Product could not be tracked"
  /tracking/{productId}/info:
    post:
      tags:
      - "Tracking"
      summary: "Information about an item being tracked by the user defined by the login token"
      description: ""
      operationId: "getTrackedProductInfo"
      consumes: 
      - "application/json"
      produces: 
      - "application/json"
      parameters:
      - name: "productId"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NewProductRequest"
      responses:
        200:
          description: "Successful"
          schema:
            $ref: "#/definitions/ProductItem"
        401:
          description: "Invalid/expired login token"
        405:
          description: "Product does not exist"
  /tracking/{productId}:
    post:
      tags: 
      - "Tracking"
      summary: "Update a tracked product"
      description: ""
      operationId: "updateProductItem"
      parameters: 
      - name: "productId"
        in: "path"
        required: true
        description: "url of the product to update"
        type: "string"
      - name: "price"
        description: "the new price value as a float/double"
        in: "query"
        required: true
        type: "number"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NewProductRequest"
      responses:
        200:
          description: "Successful"
          schema:
            $ref: "#/definitions/ProductItem"
        401:
          description: "Invalid/expired login token"
        405:
          description: "Could not update product item"

    delete:
      tags: 
      - "Tracking"
      summary: "Delete a product from the user's wishlist"
      description: ""
      operationId: "deleteTrackedProduct"
      consumes: 
      - "application/json"
      produces: 
      - "application/json"
      parameters: 
      - name: "productId"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NewProductRequest"
      responses:
        200:
          description: "Successful"
          schema:
            $ref: "#/definitions/ProductItem"
        401:
          description: "Invalid/expired login token"
        405:
          description: "Product could not be tracked"
  /tracking/batch:
    post:
      tags: 
      - "Tracking"
      summary: "Runs the batch service"
      description: "Privileged procedure. Requires an admin token."
      operationId: "executeBatch"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/LoginToken"
      responses:
        200:
          description: "Successful"
        401:
          description: "Invalid/expired login token"
definitions:
  Username:
    type: "string"
    minLength: 8
    maxLength: 32
  ProductId:
    type: "integer"
...
  LoginToken:
    type: "object"
    properties:
      username:
        $ref: "#/definitions/Username"
      sessionToken:
        type: "string"
        minLength: 128
        maxLength: 256
      expiryTime:
        type: "string"
        format: "date-time"
  ProductItem:
    type: "object"
    properties:
      productId:
        $ref: "#/definitions/ProductId"
      productName:
        type: "string"
      currentPrice:
        type: "number"
        format: "float"
      trackedPrice:
        type: "number"
        format: "float"
      trackedTime:
        type: "string"
        format: "date-time"
      vendor:
        type: "string"
      url:
        type: "string"
        format: "uri"
  NewProductRequest:
    type: "object"
    properties:
      url:
        type: "string"
        format: "uri"
      loginToken:
        $ref: "#/definitions/LoginToken"
...